# Remove render_resource_wrapper