--- a/crates/bevy_render/src/render_resource/bind_group_layout.rs
+++ b/crates/bevy_render/src/render_resource/bind_group_layout.rs
@@ -1,13 +1,14 @@
-use crate::{define_atomic_id, render_resource::resource_macros::*};
+use crate::define_atomic_id;
+use crate::renderer::WgpuWrapper;
+use alloc::sync::Arc;
 use core::ops::Deref;
 
 define_atomic_id!(BindGroupLayoutId);
-render_resource_wrapper!(ErasedBindGroupLayout, wgpu::BindGroupLayout);
 
 #[derive(Clone, Debug)]
 pub struct BindGroupLayout {
     id: BindGroupLayoutId,
-    value: ErasedBindGroupLayout,
+    value: Arc<WgpuWrapper<wgpu::BindGroupLayout>>,
 }
 
 impl PartialEq for BindGroupLayout {
@@ -32,7 +33,7 @@ impl From<wgpu::BindGroupLayout> for BindGroupLayout {
     fn from(value: wgpu::BindGroupLayout) -> Self {
         BindGroupLayout {
             id: BindGroupLayoutId::new(),
-            value: ErasedBindGroupLayout::new(value),
+            value: Arc::new(WgpuWrapper::new(value)),
         }
     }
 }