--- a/crates/bevy_render/src/render_resource/texture.rs
+++ b/crates/bevy_render/src/render_resource/texture.rs
@@ -1,10 +1,9 @@
 use crate::define_atomic_id;
+use crate::renderer::WgpuWrapper;
+use alloc::sync::Arc;
 use core::ops::Deref;
 
-use crate::render_resource::resource_macros::*;
-
 define_atomic_id!(TextureId);
-render_resource_wrapper!(ErasedTexture, wgpu::Texture);
 
 /// A GPU-accessible texture.
 ///
@@ -13,7 +12,7 @@ render_resource_wrapper!(ErasedTexture, wgpu::Texture);
 #[derive(Clone, Debug)]
 pub struct Texture {
     id: TextureId,
-    value: ErasedTexture,
+    value: Arc<WgpuWrapper<wgpu::Texture>>,
 }
 
 impl Texture {
@@ -33,7 +32,7 @@ impl From<wgpu::Texture> for Texture {
     fn from(value: wgpu::Texture) -> Self {
         Texture {
             id: TextureId::new(),
-            value: ErasedTexture::new(value),
+            value: Arc::new(WgpuWrapper::new(value)),
         }
     }
 }
@@ -48,23 +47,23 @@ impl Deref for Texture {
 }
 
 define_atomic_id!(TextureViewId);
-render_resource_wrapper!(ErasedTextureView, wgpu::TextureView);
-render_resource_wrapper!(ErasedSurfaceTexture, wgpu::SurfaceTexture);
 
 /// Describes a [`Texture`] with its associated metadata required by a pipeline or [`BindGroup`](super::BindGroup).
 #[derive(Clone, Debug)]
 pub struct TextureView {
     id: TextureViewId,
-    value: ErasedTextureView,
+    value: Arc<WgpuWrapper<wgpu::TextureView>>,
 }
 
 pub struct SurfaceTexture {
-    value: ErasedSurfaceTexture,
+    value: Arc<WgpuWrapper<wgpu::SurfaceTexture>>,
 }
 
 impl SurfaceTexture {
     pub fn try_unwrap(self) -> Option<wgpu::SurfaceTexture> {
-        self.value.try_unwrap()
+        Arc::try_unwrap(self.value)
+            .map(WgpuWrapper::into_inner)
+            .ok()
     }
 }
 
@@ -80,7 +79,7 @@ impl From<wgpu::TextureView> for TextureView {
     fn from(value: wgpu::TextureView) -> Self {
         TextureView {
             id: TextureViewId::new(),
-            value: ErasedTextureView::new(value),
+            value: Arc::new(WgpuWrapper::new(value)),
         }
     }
 }
@@ -88,7 +87,7 @@ impl From<wgpu::TextureView> for TextureView {
 impl From<wgpu::SurfaceTexture> for SurfaceTexture {
     fn from(value: wgpu::SurfaceTexture) -> Self {
         SurfaceTexture {
-            value: ErasedSurfaceTexture::new(value),
+            value: Arc::new(WgpuWrapper::new(value)),
         }
     }
 }
@@ -112,7 +111,6 @@ impl Deref for SurfaceTexture {
 }
 
 define_atomic_id!(SamplerId);
-render_resource_wrapper!(ErasedSampler, wgpu::Sampler);
 
 /// A Sampler defines how a pipeline will sample from a [`TextureView`].
 /// They define image filters (including anisotropy) and address (wrapping) modes, among other things.
@@ -122,7 +120,7 @@ render_resource_wrapper!(ErasedSampler, wgpu::Sampler);
 #[derive(Clone, Debug)]
 pub struct Sampler {
     id: SamplerId,
-    value: ErasedSampler,
+    value: Arc<WgpuWrapper<wgpu::Sampler>>,
 }
 
 impl Sampler {
@@ -137,7 +135,7 @@ impl From<wgpu::Sampler> for Sampler {
     fn from(value: wgpu::Sampler) -> Self {
         Sampler {
             id: SamplerId::new(),
-            value: ErasedSampler::new(value),
+            value: Arc::new(WgpuWrapper::new(value)),
         }
     }
 }