@tychedelia 2024-09-26
> Looks like CI / build-wasm-atomics doesn't like this change.

I think (?) we need to add `WgpuWrapper` around a bunch of these, which is gross but what I expected from removing the macro.