@bes 2024-09-26
It was a bit of an ick, but i wrapped everything in `WgpuWrapper` now...