@bes 2024-09-26
> since it seems like extra functionality (beyond the compile time optimizations) have been added to the `render_resource_wrapper`/`ErasedXXX` types, i'm wondering if we should just remove the debug version and continue to use the release version. or maybe we could wrap up the functionality into a single struct which encapsulates the current macro's release-mode behaviour.
> 
> i don't think removing the compile-time optimizations should be resulting in any call-site changes though...

This will work for everything except `wgpu::Device` which I need to be shared between the application and Bevy. Basically what you're asking for is what was in the PR to begin with?

But I'm wondering what extra functionality you mean? Deref, or the atomics stuff, or both?