@mockersf 2024-09-26
Could you check the compilation duration with this PR? The erased types were introduce for compilation speed, see https://github.com/bevyengine/bevy/pull/5950