@robtfm 2024-09-26
> Basically what you're asking for is what was in the PR to begin with, I think?

maybe? sorry i'm not sure what was in the pr to begin with (i looked when it only addressed the `RenderDevice` but not since then).

> what extra functionality you mean? Deref, or the atomics stuff, or both?

both - i think the call sites should all be unchanged. The macro can (i think) be replaced with a struct that encapsulates the extra behaviour that's been added on top of the compile time optimizations (which were only ever in the debug version of the macro).