@bes 2024-09-26
> both - i think the call sites should all be unchanged. The macro can (i think) be replaced with a struct that encapsulates the extra behaviour that's been added on top of the compile time optimizations (which were only ever in the debug version of the macro).

Any external user that depends on Bevy and wants to use a shared `wgpu::Device` would need to use this new Bevy-defined struct instead of `WgpuWrapper`, which doesn't change much, I think? Except that we can implement `Deref` for it and it doesn't change the call sites.

Do you mean something like this (just a sketch):

```rust
struct BevyArc<T> where T: Send + Sync {
    inner: Arc<send_wrapper::SendWrapper<T>>,
}

impl<T: Send + Sync> Deref for BevyArc { //...
```