@bes 2024-09-26
It'a bit depressing that we can't just use an `Arc`, but I'm also not familiar with the problem `send_wrapper` solves for Bevy, so I am keeping an open mind here.