@tychedelia 2024-09-26
> It'a bit depressing that we can't just use an `Arc`, but I'm also not familiar with the problem `send_wrapper` solves for Bevy, so I am keeping an open mind here.

I'm not familiar with the history here, I also agree it's a bit unfortunate we can't just use the arc'd types as well as `wgpu`'s ids. But there's backstory here I'm sure.