@bes 2024-09-27
I would love some guidance on what to do here. I just wanted a slightly better API for `RenderCreation:: Manual` 😅