@bes 2024-09-27
> then we just need to run compile timings - are you happy to do that or shall i do it?

I can do it if you teach me how to, but I also think that you doing it yourself might lead to better results. If you don't have an M-series mac on hand, I can do it for mac.