@robtfm 2024-09-27
It’s just `cargo build --timings`, check before and after. If the issue persists (or rather is back, since I did check it was gone a few months ago) you’ll see a ~30% slowdown in debug builds. Release would be unaffected since the macro was only type-erasing in debug. 

#5950 had a particularly bad userspace example in the top comment that would get like 95% worse as well. 

I’ll run windows tomorrow if nobody else does it in the meantime