@bes 2024-09-27
MacBook Pro M2 Max - 32GB RAM

Before my changes - commit 5fcbdc137a6e3e636aef3e5fd9190b0bc6ec2887

```shell
$ cargo clean
     Removed 148120 files, 123.9GiB total
$ cargo build --timings
...
Finished `dev` profile [unoptimized + debuginfo] target(s) in 47.31s
```

<img width="403" alt="image" src="https://github.com/user-attachments/assets/1ee6c583-7c6b-415c-afc3-1f5700ac9d65">

<img width="1051" alt="image" src="https://github.com/user-attachments/assets/88c5a186-e8e0-4868-ac0e-be69ede9b2dc">