@bes 2024-09-27
MacBook Pro M2 Max - 32GB RAM

After my changes - commit 533fa23ae1f675b70df66a36b7c564201a96c527

```shell
$ cargo clean
     Removed 19794 files, 6.0GiB total
$ cargo build --timings
...
Finished `dev` profile [unoptimized + debuginfo] target(s) in 43.66s
```

<img width="402" alt="image" src="https://github.com/user-attachments/assets/5329a48b-f553-44c7-bd09-4c7949852b5e">

<img width="967" alt="image" src="https://github.com/user-attachments/assets/acc6a32f-9460-4b53-bacd-64943b755316">