@bes 2024-09-27
For release (`cargo build --release --timings`) the numbers are:
Before - Finished `release` profile [optimized] target(s) in 1m 33s
After - Finished `release` profile [optimized] target(s) in 1m 32s