@robtfm 2024-09-27
Looks like the wasm issue is back. I think that’s because `WgpuWrapper` derives `Deref` even though its contents differ by platform. It should probably have a manual impl with target=T in the wasm case, though I don’t know what the knock on effects would be.