@bes 2024-09-28
> Looks like the wasm issue is back. I think that’s because `WgpuWrapper` derives `Deref` even though its contents differ by platform. It should probably have a manual impl with target=T in the wasm case, though I don’t know what the knock on effects would be.

I think the fix is easy, go back to the coersion i used before: `layout: layout.as_ref().map(|layout| -> &PipelineLayout { layout }),`. If you think it's unclear, I can add comments above to explain.

As I wrote in the review comment:

> Your suggestion doesn't work, since WgpuWrapper has a different type signature between wasm32 and everything else, so your specific number of derefs &** only works for the non-wasm32-version.
>
> My version works reliably for both, because the compiler will insert the correct number of derefs at compile time.
>
> What do you want me to do?