@robtfm 2024-09-28
Yes go ahead, your original way makes sense. Sorry!