@bes 2024-09-28
> Yes go ahead, your original way makes sense. Sorry!

Great! Updated! Let me know how to proceed.