@cart 2025-03-12
Hmm this changeset costs 10 seconds on current main for me on linux (1 min 2 seconds vs 52 seconds). Perhaps this optimization only comes into play on linux?