@robtfm 2025-03-12
rerunning on windows with rustc 1.85
pr is 1m31s, df23b937 (the prior commit) is 1m34s

so win still looks good. pretty surprising, i thought this was a compiler trait-bound rabbit hole which shouldn't be platform dependent at all...