@mockersf 2025-03-12
I also see a slowdown with this PR on macos on a m4, a bit on a smaller scale that what Cart is reporting, but with more variation (running on a laptop, and efficiency/performance cores)

To run a good benchmark for this kind of things, you should use something like [`hyperfine`](https://github.com/sharkdp/hyperfine).

```sh
git checkout 72aaa41603d28d5d6ed6906f57f6c596ba3868c3 # this PR
hyperfine --prepare 'sleep 1; cargo clean' 'RUSTC_WRAPPER= cargo build --release'
git checkout HEAD~1
hyperfine --prepare 'sleep 1; cargo clean' 'RUSTC_WRAPPER= cargo build --release'
```

This PR:
```
  Time (mean ± σ):     81.554 s ±  4.136 s    [User: 1059.526 s, System: 38.421 s]
  Range (min … max):   73.708 s … 85.835 s    10 runs
```

`main` just before:
```
  Time (mean ± σ):     77.674 s ±  7.937 s    [User: 1019.559 s, System: 37.408 s]
  Range (min … max):   62.272 s … 83.375 s    10 runs
```